alert_generation/requirements.txt
pyparsing<3,>=2.0.2
pyrebase4==4.5.0
sendgrid==6.9.7
